requirements.txt
openai
pandas
pyarrow
pyyaml
python-dotenv
bert-score
sacrebleu
scipy
tqdm
tenacity
matplotlib
seaborn
transformers
torch
sacremoses
accelerate
sentencepiece
protobuf
